Create base class for EvaluateOperation and OptimizeOperation.
